# auth-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Documentation for NeoCentral Auth Service",
        "title": "NeoCentral Auth Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@neocentral.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8001",
    "basePath": "/",
    "paths": {
...
        "/profile/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves the profile information of the currently authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Get My Profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/neocentral-go_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/neocentral-go_auth-service_internal_dto.UserProfileResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/neocentral-go_pkg_response.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "neocentral-go_auth-service_internal_dto.LecturerResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "scienceGroupId": {
                    "type": "string"
                }
            }
        },
...
        "neocentral-go_auth-service_internal_dto.RoleResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "neocentral-go_auth-service_internal_dto.StudentResponse": {
            "type": "object",
            "properties": {
                "enrollmentYear": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "sksCompleted": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "neocentral-go_auth-service_internal_dto.UserProfileResponse": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "identityNumber": {
                    "type": "string"
                },
                "identityType": {
                    "type": "string"
                },
                "isVerified": {
                    "type": "boolean"
                },
                "lecturer": {
                    "$ref": "#/definitions/neocentral-go_auth-service_internal_dto.LecturerResponse"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/neocentral-go_auth-service_internal_dto.RoleResponse"
                    }
                },
                "student": {
                    "$ref": "#/definitions/neocentral-go_auth-service_internal_dto.StudentResponse"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "neocentral-go_pkg_response.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}